refactor: restructure index files
