--- src/elements/listitem/listitem.ts
@@ -1,8 +1,0 @@
-import {LitElement} from "lit";
-
-export default class extends LitElement {
-  protected override createRenderRoot() {
-    // No shadow DOM.
-    return this;
-  }
-}
