+++ src/elements/locale/index.ts
@@ -1,0 +1,17 @@
+import HTMLIntlLocaleElement from './locale';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'intl-locale': HTMLIntlLocaleElement;
+  }
+}
+
+export function defineIntlLocaleElements() {
+  customElements.define('intl-locale', HTMLIntlLocaleElement);
+}
+
+export {
+  HTMLIntlLocaleElement,
+}
+
+
